+++ CHANGELOG.md
@@ -17,7 +17,7 @@ We use the following categories for changes:
 ## [Unreleased]
 
 ### Added
-
+- added get client token command [#sc-11670]
 ### Changed
 
 ### Fixed