+++ CHANGELOG.md
@@ -17,7 +17,9 @@ We use the following categories for changes:
 ## [Unreleased]
 
 ### Added
+- added get datasources API key command [#sc-11671]
 - added get client token command [#sc-11670]
+
 ### Changed
 
 ### Fixed